requirements.txt
asgiref==3.4.1
Django==4.0
django-ckeditor==6.2.0
django-crispy-forms==1.13.0
django-filter==21.1
django-js-asset==1.2.2
django-quill-editor==0.1.29
gunicorn==20.1.0
Pillow==8.4.0
sqlparse==0.4.2
tzdata==2021.5
whitenoise==5.3.0
